+++ scrapypuppeteer/middleware.py
@@ -31,17 +31,8 @@ from scrapypuppeteer.response import (
 )
 from scrapypuppeteer.request import ActionRequest, PuppeteerRequest, CloseContextRequest
 
-
-
-
-from scrapypuppeteer.service_browser_manager import ServiceBrowserManager
-
-from scrapypuppeteer.local_browser_manager import LocalBrowserManager
-
-
-
-
-    
+from scrapypuppeteer.browser_managers.local_browser_manager import LocalBrowserManager
+from scrapypuppeteer.browser_managers.service_browser_manager import ServiceBrowserManager
 
 class PuppeteerServiceDownloaderMiddleware:
     """
